pyproject.toml
[project]
name = "visionlog"
version = "0.2.0"
description = "The static St. Peter — vision, goals, guardrails, SOPs, ADRs for AI agent governance"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Daniel Shanklin", email = "daniel@eidosagi.com" },
]
keywords = ["mcp", "vision", "governance", "guardrails", "agent-tools"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
visionlog = "visionlog_md.server:main"

[project.entry-points."mcp"]
visionlog = "visionlog_md.server:mcp"

[project.urls]
Homepage = "https://github.com/eidos-agi/visionlog.md"
Repository = "https://github.com/eidos-agi/visionlog.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["visionlog_md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
